requirements.txt
fastapi==0.95.2
uvicorn==0.22.0
texthero==1.1.0
psycopg2==2.9.5
sqlalchemy==2.0.17
pydantic==1.10.8
pandas==2.0.2
numpy==1.24.3
scikit-learn==1.2.2
catboos==1.1.1